chore: Move `Period` to time package
