+++ src/main/java/meditracker/dailymedication/DailyMedicationManager.java
@@ -4,6 +4,7 @@ import meditracker.exception.FileReadWriteException;
 import meditracker.medication.Medication;
 import meditracker.medication.MedicationManager;
 import meditracker.storage.FileReaderWriter;
+import meditracker.time.Period;
 import meditracker.ui.Ui;
 
 import java.time.LocalDate;
