--- src/main/java/meditracker/dailymedication/Period.java
@@ -1,7 +1,0 @@
-package meditracker.dailymedication;
-
-public enum Period {
-    MORNING,
-    AFTERNOON,
-    EVENING;
-}
